requirements.txt
langchain
langchain-community
langchain-openai
numpy
openai
tiktoken
python-dotenv
slack-bolt
flask
